size_t changes

+++ edbee-lib/edbee/models/change.h
@@ -97,8 +97,8 @@ public:
     virtual void flatten();
 
     virtual void giveChange(TextDocument* doc, Change* change);
-    virtual Change* at(int idx);
-    virtual Change* take(int idx);
+    virtual Change* at(size_t idx);
+    virtual Change* take(size_t idx);
     virtual size_t size();
     virtual void clear(bool performDelete=true);
     Change* last();
